requirements.txt
streamlit
llama-index
llama-index-llms-nvidia
llama-index-embeddings-nvidia
llama-parse
python-dotenv
nest-asyncio
uvicorn
fastapi
pydantic

